moved wheel picker into desktop and android specific implementations
